Refactored event listeners
