+++ index.html
@@ -55,9 +55,6 @@
               <input type="name" class="form-control" id="stateName" placeholder="Enter Name">
             </div>
             <div class="form-group">
-              <input type="name" class="form-control" id="stateYear" placeholder="Enter year 2000-2014">
-            </div>
-            <div class="form-group">
               <input type="name" class="form-control" id="stateId" placeholder="Enter 2 letter state abbreviation">
             </div>
             <div class="form-check">
@@ -77,7 +74,6 @@
     </div>
   </div>
   </ul>
-
   </nav>
 
   <main role="main" class="col-sm-9 ml-sm-auto col-md-10 pt-3" id="results">
@@ -88,16 +84,8 @@
   </main>
 
 </body>
-
-
-
-
-
-
-
-
-
-  <script src="main.bundle.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.5.0/Chart.min.js"></script>
+<script src="main.bundle.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.5.0/Chart.min.js"></script>
 </body>
+
 </html>
