+++ lib/requests/requests.js
@@ -1,14 +1,31 @@
 const $ = require('jquery')
-const url = 'http://localhost:3000/api/v1/'
+const responses = require('../response/response.js')
 
-const getResponse = (path) => {
-  $.get(url + path , {method: 'GET'})
-    .then( response => console.log(response))
+const getGraph = (path) => {
+  console.log("something")
+  $.get(path , {method: 'GET'})
+    .then( response => responses.drawGraph(response))
+    .catch( error => {console.log( { error })
+  })
+}
+
+const getTable = (path, id1, id2) => {
+  $.get(path , {method: 'GET'})
+    .then( response => responses.drawTable(response, id1, id2))
+    .catch( error => {console.log( { error })
+  })
+}
+
+const getStateTable = (path) => {
+  $.get(path , {method: 'GET'})
+    .then( response => responses.drawStateTable(response))
     .catch( error => {console.log( { error })
   })
 }
 
 
 module.exports = {
-  getResponse
+  getGraph,
+  getTable,
+  getStateTable
 }
